pyproject.toml
[project]
name = "cheatpug"
version = "0.1.0"
description = ""
authors = [
    {name = "Sasha Tsuzanovsky",email = "a.tsuzanovsky@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "telethon (>=1.40.0,<2.0.0)",
    "telebot (>=0.0.5,<0.0.6)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai (>=1.102.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
